cnc_geometry: Calculator headered and fixed
